+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Development tooling with `uv` (#85)
 - Support for Django 5.2, Wagtail 7.0 and 7.1. (#85)
+- Add extra context to `get_newsletter_html` (#88)
 
 ### Removed